backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.25.1
azure-mgmt-compute==37.1.0
azure-mgmt-core==1.6.0
azure-mgmt-costmanagement==4.0.1
azure-monitor-query==2.0.0
billiard==4.2.4
boto3==1.34.8
botocore==1.34.162
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmdstanpy==1.3.0
contourpy==1.3.3
cron_descriptor==2.0.6
cryptography==46.0.3
cycler==0.12.1
Django==4.2.8
django-celery-beat==2.8.1
django-celery-results==2.5.1
django-timezone-field==7.2.1
djangorestframework==3.14.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-bigquery==3.13.0
google-cloud-compute==1.15.0
google-cloud-core==2.5.0
google-cloud-monitoring==2.16.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
holidays==0.89
idna==3.11
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
networkx==3.6.1
numpy==1.26.2
oauthlib==3.3.1
packaging==25.0
pandas==2.1.4
pillow==12.1.0
prompt_toolkit==3.0.52
prophet==1.1.5
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
pyparsing==3.3.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
reportlab==4.4.9
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.10.4
scikit-learn==1.3.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
stanio==0.5.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.1.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
